split path builder from pen
