+++ libs/content/workspace/src/tab/svg_editor/mod.rs
@@ -3,6 +3,7 @@ mod eraser;
 mod gesture_handler;
 mod history;
 mod parser;
+mod path_builder;
 mod pen;
 mod renderer;
 mod selection;
@@ -19,7 +20,7 @@ pub use history::InsertElement;
 use lb_rs::Uuid;
 pub use parser::Buffer;
 use parser::DiffState;
-pub use pen::CubicBezBuilder;
+pub use path_builder::PathBuilder;
 pub use pen::Pen;
 use renderer::Renderer;
 use resvg::usvg::ImageKind;
